pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawlock"
version = "1.0.0"
description = "Comprehensive security scanner, red-teamer & hardening toolkit for Claw-based AI agent deployments"
readme = "README.md"
license = {text = "Apache-2.0 OR MIT"}
requires-python = ">=3.9"
authors = [{name = "g1at"}]
keywords = ["security", "ai-agent", "openclaw", "zeroclaw", "claude-code", "mcp", "red-team", "hardening", "llm-security", "prompt-injection", "supply-chain", "tool-poisoning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]
dependencies = ["typer>=0.12", "rich>=13.0", "pyyaml>=6.0", "httpx>=0.25"]

[project.optional-dependencies]
redteam = ["promptfoo>=0.80"]
dev = ["pytest>=7.0", "ruff>=0.3", "mypy>=1.8"]

[project.scripts]
clawlock = "clawlock.__main__:app"

[project.urls]
Homepage = "https://github.com/g1at/clawlock"
Repository = "https://github.com/g1at/clawlock"

[tool.setuptools.packages.find]
include = ["clawlock*"]
exclude = ["tests*", "rules*", "skill*"]
